Keg.find_some_installed_dependents: move to installed_dependents.rb
