+++ Library/Homebrew/uninstall.rb
@@ -1,7 +1,7 @@
 # typed: true
 # frozen_string_literal: true
 
-require "keg"
+require "installed_dependents"
 
 module Homebrew
   # Helper module for uninstalling kegs.
@@ -108,7 +108,7 @@ module Homebrew
     end
 
     def check_for_dependents(kegs, casks: [], named_args: [])
-      return false unless result = Keg.find_some_installed_dependents(kegs, casks: casks)
+      return false unless result = InstalledDependents.find_some_installed_dependents(kegs, casks: casks)
 
       if Homebrew::EnvConfig.developer?
         DeveloperDependentsMessage.new(*result, named_args: named_args).output
